requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clr-loader==0.2.5
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.12.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
line-bot-sdk==2.4.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
multidict==6.0.4
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pythonnet==3.0.1
pytweening==1.0.7
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.28.2
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
zipp==3.15.0
